Split NDJSON row parsing
