+++ jetro-core/src/io/ndjson.rs
@@ -16,6 +16,7 @@ use super::ndjson_frame::{frame_payload, FramePayload, NdjsonRowFrame};
 use super::ndjson_hint::{
     NdjsonHintAccessPlan, NdjsonHintConfig, NdjsonHintDecision, NdjsonHintState,
 };
+pub(super) use super::ndjson_row::{collect_row_val, parse_row, row_eval_error, row_parse_error};
 use super::ndjson_rows::{
     ndjson_rows_file_plan, ndjson_rows_stream_plan, ndjson_rows_subquery_plan, NdjsonRowsFilePlan,
 };
@@ -3562,48 +3563,6 @@ fn write_control_escape<W: Write>(writer: &mut W, byte: u8) -> Result<(), JetroE
         HEX[(byte & 0x0f) as usize],
     ])?;
     Ok(())
-}
-
-pub(super) fn collect_row_val(
-    engine: &JetroEngine,
-    document: &Jetro,
-    plan: &crate::ir::physical::QueryPlan,
-    line_no: u64,
-) -> Result<Val, JetroEngineError> {
-    engine
-        .collect_prepared_val(document, plan)
-        .map_err(|err| row_eval_error(line_no, err))
-}
-
-pub(super) fn parse_row(
-    engine: &JetroEngine,
-    line_no: u64,
-    row: Vec<u8>,
-) -> Result<Jetro, JetroEngineError> {
-    engine
-        .parse_bytes_lazy(row)
-        .map_err(|err| row_parse_error(line_no, err))
-}
-
-pub(super) fn row_parse_error(line_no: u64, err: JetroEngineError) -> JetroEngineError {
-    match err {
-        JetroEngineError::Json(source) => RowError::InvalidJson { line_no, source }.into(),
-        JetroEngineError::Eval(eval) => RowError::InvalidJsonMessage {
-            line_no,
-            message: eval.to_string(),
-        }
-        .into(),
-        other => other,
-    }
-}
-
-pub(super) fn row_eval_error(line_no: u64, err: crate::EvalError) -> JetroEngineError {
-    let message = err.0;
-    if message.starts_with("Invalid JSON:") {
-        RowError::InvalidJsonMessage { line_no, message }.into()
-    } else {
-        crate::EvalError(message).into()
-    }
 }
 
 fn trim_line_ending(buf: &mut Vec<u8>) {
